emulation: fix renode compatibility for peripherals
Many of the devicse have changed from `uint` to `ulong`, so forward-port
the peripherals to work with the new format.

+++ emulation/peripherals/LiteX_Timer_32.cs
@@ -113,8 +113,8 @@ namespace Antmicro.Renode.Peripherals.Timers.Betrusted
         private IFlagRegisterField irqPending;
 
         private uint latchedValue;
-        private uint loadValue;
-        private uint reloadValue;
+        private ulong loadValue;
+        private ulong reloadValue;
 
         private readonly LimitTimer innerTimer;
